Add files via upload
I am uploading manually because I was unable to push to the remote, potentially due to issues with my SSH key. Note that I used the exact file path when directing the program to the files because the computer was unable to locate the correct files when I used a shorter path location.